# api/openapi.yaml
openapi: 3.1.0
info:
  title: doenerstag
  version: "1"
  summary: Coordinate food orders for a group.
  description: |
    doenerstag coordinates food orders for a group. It does not take payments
    and does not place orders with restaurants.

    This document describes every endpoint the server registers. A test walks
    the routes of the production router and fails when one of them is missing
    from here, and fails the other way too when this document describes a path
    nothing serves — so the two cannot drift apart.

    Reading is mostly open. An unauthenticated visitor sees restaurants, menus,
    the list of orders and each order's item count, but never who ordered what:
    the split is enforced in the handlers and is described at each endpoint it
    affects.

    Collections are wrapped in an object with one plural key rather than being
    returned as a bare array — `{"restaurants": [...]}` — which leaves room to
    add a field beside the list without changing the type of the response.
  license:
    name: BSD-3-Clause
    identifier: BSD-3-Clause
servers:
  - url: /api/v1
    description: This installation.
...
security:
  - session: []
  - token: []
paths:
...
  /restaurants/{id}/menu-items:
    parameters:
      - $ref: "#/components/parameters/RestaurantId"
    get:
      tags: [menu]
      operationId: listMenuItems
      summary: The menu, in a printed menu's order.
      description: |
        Ordered by the restaurant's own item number, numerically where it is a
        number so that 2 comes before 10, then by name. Items without a number
        come last.

        The filters combine with AND across parameter names and OR within one:
        `?tag=vegan&tag=spicy&exclude_allergen=nuts` means "vegan or spicy, and
        not containing nuts".
      security: []
      parameters:
        - name: category
          in: query
          description: Restrict to a category.
          schema: { type: string, format: uuid }
        - name: tag
          in: query
          description: Item carries any of these tags.
          schema: { type: array, items: { type: string } }
        - name: exclude_allergen
          in: query
          description: Item does not carry this allergen.
          schema: { type: array, items: { type: string } }
        - name: exclude_additive
          in: query
          description: Item does not carry this additive.
          schema: { type: array, items: { type: string } }
        - name: available
          in: query
          description: Only items that can be ordered.
          schema: { type: boolean }
      responses:
        "200":
          description: The matching items.
          content:
            application/json:
              schema:
                type: object
                required: [menu_items]
                properties:
                  menu_items:
                    type: array
                    items:
                      $ref: "#/components/schemas/MenuItem"
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags: [menu]
      operationId: createMenuItem
      summary: Add a menu item.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MenuItemRequest"
      responses:
        "201":
          description: The new item.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MenuItem"
        "400":
          $ref: "#/components/responses/Validation"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    session:
      type: apiKey
      in: cookie
      name: doener_session
      description: |
        A JWT backed by a server-side session row. State-changing requests must
        also send the CSRF token from the doener_csrf cookie in the
        X-CSRF-Token header.
    token:
      type: http
      scheme: bearer
      description: |
        An API token for third-party and scripted access. Not subject to the
        CSRF check, because it carries no ambient browser credentials.
  parameters:
...
    RestaurantId:
      name: id
      in: path
      required: true
      description: The restaurant's id.
      schema: { type: string, format: uuid }
...
  responses:
...
    Validation:
      description: The request was refused. Errors 1000 to 1999.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Authentication is required, or the session has ended. Errors 2000 to 2999.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: No such thing. Error 4000.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Conflict:
      description: The request conflicts with the current state. Errors 4001 to 4999.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Error:
      type: object
      required: [error]
      description: |
        Every failure carries this shape. The code is an internal, stable
        number the frontend keys its translations on; the message is
        developer-facing English and is never the primary way a user learns
        what went wrong.

        The request_id matches the X-Request-Id response header and the log
        line, which is what makes a report of "it went wrong just now"
        traceable without the response ever carrying SQL text or a stack trace.
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: integer
              description: The internal error number. See docs/04_api.md.
            message:
              type: string
            field:
              type: [string, "null"]
              description: The offending request field, for validation errors.
            request_id:
              type: string
...
    Tag:
      type: object
      required: [id, code, name, sort_order]
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name:
          type: string
          description: |
            Only meaningful for a tag somebody invented. A seeded tag is
            translated from its code.
        sort_order: { type: integer }
    Classification:
      type: object
      required: [id, code, reference, sort_order]
      description: An allergen or an additive.
      properties:
        id: { type: string, format: uuid }
        code: { type: string, examples: ["gluten"] }
        reference:
          type: string
          description: The regulatory or conventional number.
        sort_order: { type: integer }
...
    MenuItem:
      type: object
      required:
        - id
        - restaurant_id
        - external_id
        - name
        - description
        - price_cents
        - available
        - tags
        - allergens
        - additives
      properties:
        id: { type: string, format: uuid }
        restaurant_id: { type: string, format: uuid }
        category_id: { type: [string, "null"], format: uuid }
        external_id:
          type: string
          description: The restaurant's own item number, if it has one.
        name: { type: string }
        description: { type: string }
        image_id: { type: [string, "null"], format: uuid }
        price_cents: { type: integer, format: int64 }
        available: { type: boolean }
        tags:
          type: array
          items:
            $ref: "#/components/schemas/Tag"
        allergens:
          type: array
          items:
            $ref: "#/components/schemas/Classification"
        additives:
          type: array
          items:
            $ref: "#/components/schemas/Classification"
        modifications:
          type: array
          description: Present when one item is fetched, absent from the list.
          items:
            $ref: "#/components/schemas/Modification"
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    MenuItemRequest:
      type: object
      required: [name, price_cents]
      properties:
        category_id: { type: [string, "null"], format: uuid }
        external_id: { type: string }
        name: { type: string }
        description: { type: string }
        image_id: { type: [string, "null"], format: uuid }
        price_cents: { type: integer, format: int64 }
        available: { type: boolean }
        tag_ids: { type: array, items: { type: string, format: uuid } }
        allergen_ids: { type: array, items: { type: string, format: uuid } }
        additive_ids: { type: array, items: { type: string, format: uuid } }
...
    Modification:
      type: object
      required: [id, name, price_delta_cents, sort_order]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        price_delta_cents:
          type: integer
          format: int64
          description: May be negative, for an option that takes something away.
        sort_order: { type: integer }
...